Basic example of autogeneration of blinkies
for the purpose of running a mini blink, there's no actual different in
ld scripts needed at all, we just need one that supports our standard
startup, and declares enough ram/flash to be functional, but small
enough that it should cover just about any of the boards we can find.

Makefile's not ideal, but it _does_ correctly allow parallel builds at
least.